08_Practice/requirements.txt
mysql-connector-python==9.0.0
faker==30.3.0
